move state definition details
